Fixed the Surefire plugin's configuration to run the tests correctly.
Changed the Maven Surefire plugin configuration to not run the integration tests in parallel, as this causes some of them to fail. They now execute in the correct order.
