+++ StudentManagementSystem/src/test/java/com/yoanpetrov/studentmanagementsystem/integration/UserIntegrationTests.java
@@ -1,4 +1,4 @@
-package com.yoanpetrov.studentmanagementsystem.controller;
+package com.yoanpetrov.studentmanagementsystem.integration;
 
 import com.yoanpetrov.studentmanagementsystem.model.Course;
 import com.yoanpetrov.studentmanagementsystem.model.User;
@@ -14,16 +14,15 @@ import org.springframework.http.HttpStatus;
 import java.util.Arrays;
 
 import static io.restassured.RestAssured.*;
-import static io.restassured.RestAssured.given;
+import static org.junit.jupiter.api.Assertions.*;
+import static org.hamcrest.Matchers.*;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.equalTo;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class CourseIntegrationTests {
+public class UserIntegrationTests {
 
-    private static final String BASE_URI = "api/v1/courses";
+    private static final String BASE_URI = "api/v1/users";
 
     private static final User TEST_USER = User.builder() // TODO: 03-Oct-23 Put these in a Context and autowire them
             .userId(1L)
@@ -33,12 +32,13 @@ public class CourseIntegrationTests {
             .password("test")
             .role(Role.USER).build();
 
-    private static final Course UPDATED_COURSE = Course.builder()
-            .courseId(1L)
-            .name("Test")
-            .description("Updated test description")
-            .maxCapacity(10)
-            .numberOfStudents(0).build();
+    private static final User UPDATED_USER = User.builder()
+            .userId(1L)
+            .firstName("Test")
+            .lastName("Updated")
+            .email("test@test.com")
+            .password("test")
+            .role(Role.USER).build();
 
     private static final Course TEST_COURSE = Course.builder()
             .courseId(1L)
@@ -62,8 +62,8 @@ public class CourseIntegrationTests {
 
     @Order(1)
     @Test
-    public void testCreateCourse() {
-        with().body(TEST_COURSE)
+    public void testCreateUser() {
+        with().body(TEST_USER)
                 .when()
                 .contentType(ContentType.JSON)
                 .post(BASE_URI)
@@ -74,60 +74,8 @@ public class CourseIntegrationTests {
 
     @Order(2)
     @Test
-    public void testGetAllCourses() {
-        Course[] courses = get(BASE_URI).then()
-                .statusCode(HttpStatus.OK.value())
-                .extract().as(Course[].class);
-
-        assertThat(courses.length, equalTo(1));
-        assertTrue(Arrays.stream(courses).toList().contains(TEST_COURSE));
-    }
-
-    @Order(3)
-    @Test
-    public void testGetCourseById() {
-        with().get(BASE_URI + "/1").then()
-                .assertThat()
-                .body(
-                        "courseId", equalTo(1),
-                        "name", equalTo("Test"),
-                        "description", equalTo("Test description"),
-                        "maxCapacity", equalTo(10),
-                        "numberOfStudents", equalTo(0));
-    }
-
-    @Order(4)
-    @Test
-    public void testUpdateCourse() {
-        with().body(UPDATED_COURSE).when()
-                .contentType(ContentType.JSON)
-                .put(BASE_URI + "/1").then()
-                .assertThat()
-                .statusCode(HttpStatus.OK.value())
-                .and()
-                .body("description", equalTo("Updated test description"));
-    }
-
-    @Order(5)
-    @Test
-    public void testAddUserToCourse() {
-        with().body(TEST_USER)
-                .contentType(ContentType.JSON)
-                .post("api/v1/users").then()
-                .assertThat()
-                .statusCode(HttpStatus.CREATED.value());
-        with().body(TEST_USER)
-                .contentType(ContentType.JSON)
-                .post(BASE_URI + "/1/users").then()
-                .assertThat()
-                .statusCode(HttpStatus.CREATED.value());
-        TEST_COURSE.setNumberOfStudents(TEST_COURSE.getNumberOfStudents() + 1);
-    }
-
-    @Order(6)
-    @Test
-    public void testGetAllCourseUsers() {
-        User[] users = get(BASE_URI + "/1/users").then()
+    public void testGetAllUsers() {
+        User[] users = get(BASE_URI).then()
                 .statusCode(HttpStatus.OK.value())
                 .extract().as(User[].class);
 
@@ -135,24 +83,77 @@ public class CourseIntegrationTests {
         assertTrue(Arrays.stream(users).toList().contains(TEST_USER));
     }
 
+    @Order(3)
+    @Test
+    public void testGetUserById() {
+        with().get(BASE_URI + "/1").then()
+                .assertThat()
+                .body(
+                        "userId", equalTo(1),
+                        "firstName", equalTo("Test"),
+                        "lastName", equalTo("User"),
+                        "email", equalTo("test@test.com"),
+                        "email", equalTo("test@test.com"),
+                        "role", equalTo("USER"));
+    }
+
+    @Order(4)
+    @Test
+    public void testUpdateUser() {
+        with().body(UPDATED_USER).when()
+                .contentType(ContentType.JSON)
+                .put(BASE_URI + "/1").then()
+                .assertThat()
+                .statusCode(HttpStatus.OK.value())
+                .and()
+                .body("lastName", equalTo("Updated"));
+    }
+
+    @Order(5)
+    @Test
+    public void testAddCourseToUser() {
+        with().body(TEST_COURSE)
+                .contentType(ContentType.JSON)
+                .post("api/v1/courses").then()
+                .assertThat()
+                .statusCode(HttpStatus.CREATED.value());
+        with().body(TEST_COURSE)
+                .contentType(ContentType.JSON)
+                .post(BASE_URI + "/1/courses").then()
+                .assertThat()
+                .statusCode(HttpStatus.CREATED.value());
+        TEST_COURSE.setNumberOfStudents(TEST_COURSE.getNumberOfStudents() + 1);
+    }
+
+    @Order(6)
+    @Test
+    public void testGetAllUserCourses() {
+        Course[] courses = get(BASE_URI + "/1/courses").then()
+                .statusCode(HttpStatus.OK.value())
+                .extract().as(Course[].class);
+
+        assertThat(courses.length, equalTo(1));
+        assertTrue(Arrays.stream(courses).toList().contains(TEST_COURSE));
+    }
+
     @Order(7)
     @Test
-    public void testRemoveUserFromCourse() {
-        with().body(TEST_USER)
+    public void testRemoveCourseFromUser() {
+        with().body(TEST_COURSE)
                 .contentType(ContentType.JSON)
-                .delete(BASE_URI + "/1/users").then()
+                .delete(BASE_URI + "/1/courses").then()
                 .assertThat()
                 .statusCode(HttpStatus.OK.value());
-        given().get(BASE_URI + "/1/users").then()
+        given().get(BASE_URI + "/1/courses").then()
                 .assertThat()
                 .statusCode(HttpStatus.NO_CONTENT.value());
     }
 
     @Order(8)
     @Test
-    public void testRemoveCourseById() {
-        TEST_COURSE.setCourseId(2L);
-        with().body(TEST_COURSE)
+    public void testRemoveUserById() {
+        TEST_USER.setUserId(2L);
+        with().body(TEST_USER)
                 .contentType(ContentType.JSON)
                 .post(BASE_URI).then()
                 .assertThat()
@@ -164,7 +165,7 @@ public class CourseIntegrationTests {
 
     @Order(9)
     @Test
-    public void testRemoveAllCourses() {
+    public void testRemoveAllUsers() {
         given().delete(BASE_URI).then()
                 .assertThat()
                 .statusCode(HttpStatus.OK.value());
